requirements.txt
Flask==2.3.0
pymysql==1.0.2